requirements.txt
python-dateutil
pandas
matplotlib
ijson
networkx
scikit-learn
seaborn
pytest
pytest-cov
pytest-mock
